chore(squash): improve tests
